+++ e2e/src/lib.rs
@@ -4,24 +4,84 @@ mod client;
 mod fuel_node;
 #[cfg(test)]
 mod kms;
+#[cfg(test)]
+mod e2e_helpers;
+
 
 #[cfg(test)]
 mod tests {
-    use crate::fuel_node::FuelNode;
     use anyhow::Result;
-    use crate::kms::Kms;
+    use fuels::prelude::{AssetId, Provider};
+    use std::str::FromStr;
+    use fuels_accounts::aws::AwsWallet;
+    use fuels_accounts::ViewOnlyAccount;
+    use crate::e2e_helpers::{create_and_fund_kms_keys, start_fuel_node, start_kms};
 
-    #[tokio::test(flavor = "multi_thread")]
-    async fn aws_wallet() -> Result<()> {
+    #[tokio::test]
+    async fn fund_aws_wallet() -> Result<()> {
+        let kms = start_kms(false).await?;
+        let fuel_node = start_fuel_node(false).await?;
+        let kms_key = create_and_fund_kms_keys(&kms, &fuel_node).await?;
 
-        let kms = Kms::default().with_show_logs(false).start().await?;
-        let key = kms.create_key().await?;
-        let fuel_node_process = FuelNode::default().with_show_logs(false).start().await?;
-        
-        fuel_node_process.fund(key.kms_data.address, 5_000_000_000).await?;
+        std::env::set_var("AWS_ACCESS_KEY_ID", "test");
+        std::env::set_var("AWS_SECRET_ACCESS_KEY", "test");
+        std::env::set_var("AWS_REGION", "us-east-1");
+        std::env::set_var("AWS_ENDPOINT_URL", &kms_key.url);
 
+        let asset_id =
+            AssetId::from_str("f8f8b6283d7fa5b672b530cbb84fcccb4ff8dc40f8176ef4544ddb1f1952ad07")
+                .expect("AssetId to be well formed");
 
+        let provider = Provider::connect(fuel_node.url()).await?;
+        let wallet = AwsWallet::from_kms_key_id(kms_key.id, Some(provider)).await?;
+
+        let founded_coins = wallet.get_coins(asset_id).await?.first().expect("No coins found").amount;
+        assert_eq!(founded_coins, 5000000000);
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn deploy_contract() -> anyhow::Result<()> {
+        use fuels::prelude::*;
+
+        let kms = start_kms(false).await?;
+        let fuel_node = start_fuel_node(false).await?;
+        let kms_key = create_and_fund_kms_keys(&kms, &fuel_node).await?;
+
+        std::env::set_var("AWS_ACCESS_KEY_ID", "test");
+        std::env::set_var("AWS_SECRET_ACCESS_KEY", "test");
+        std::env::set_var("AWS_REGION", "us-east-1");
+        std::env::set_var("AWS_ENDPOINT_URL", &kms_key.url);
+
+        let asset_id =
+            AssetId::from_str("f8f8b6283d7fa5b672b530cbb84fcccb4ff8dc40f8176ef4544ddb1f1952ad07")
+                .expect("AssetId to be well formed");
+
+        let provider = Provider::connect(fuel_node.url()).await?;
+        let wallet = AwsWallet::from_kms_key_id(kms_key.id, Some(provider)).await?;
+
+        let founded_coins = wallet.get_coins(asset_id).await?.first().expect("No coins found").amount;
+        assert_eq!(founded_coins, 5000000000);
+
+        let contract_id = Contract::load_from(
+            "../e2e/sway/contracts/contract_test/out/release/contract_test.bin",
+            LoadConfiguration::default(),
+        )?
+            .deploy(&wallet, TxPolicies::default())
+            .await?;
+
+        println!("Contract deployed @ {contract_id}");
+
+        let founded_coins = wallet.get_coins(asset_id).await?.first().expect("No coins found").amount;
+        assert_eq!(founded_coins, 4999983198);
 
         Ok(())
     }
+
+
+
+
+
+
+
 }
