+++ packages/fuels-accounts/src/aws.rs
@@ -1,0 +1,5 @@
+mod aws_client;
+mod aws_signer;
+
+pub use aws_client::*;
+pub use aws_signer::*;
